+++ src/ecs/syscall.rs
@@ -7,6 +7,8 @@ use bevy::prelude::*;
 //standard shortcuts
 use std::marker::PhantomData;
 
+use crate::react::ReactorResult;
+
 //-------------------------------------------------------------------------------------------------------------------
 //-------------------------------------------------------------------------------------------------------------------
 
@@ -64,7 +66,7 @@ where
 /// assert_eq!(world.syscall(1u16, test_function), 2u16);
 /// ```
 ///
-pub fn syscall<I, O, S, Marker>(world: &mut World, input: <I as bevy::prelude::SystemInput>::Inner<'_>, system: S) -> O
+pub fn syscall<I, O, S, Marker>(world: &mut World, input: <I as SystemInput>::Inner<'_>, system: S) -> O
 where
     I: Send + Sync + SystemInput + 'static,
     O: Send + Sync + 'static,
@@ -80,7 +82,7 @@ where
 /// The validation function can be used to check for resources and print friendly error messages.
 pub fn syscall_with_validation<I, O, S, Marker>(
     world: &mut World,
-    input: <I as bevy::prelude::SystemInput>::Inner<'_>,
+    input: <I as SystemInput>::Inner<'_>,
     system: S,
     validation: fn(&mut World)
 ) -> O
@@ -120,11 +122,12 @@ where
 /// This is intended to wrap `Fn` systems. Do not use it if you have a `FnOnce` callback, for example when
 /// adding a one-off callback via `Command::add()`, because the input value and system will be unnecessarily cloned.
 pub fn prep_fncall<I, O, Marker>(
-    input  : <I as bevy::prelude::SystemInput>::Inner<'static>,
+    input  : <I as SystemInput>::Inner<'static>,
     system : impl IntoSystem<I, O, Marker> + Send + Sync + 'static + Clone
 ) -> impl Fn(&mut World) -> O + Send + Sync + 'static
 where
-    I: Send + Sync + SystemInput + 'static + Clone, <I as bevy::prelude::SystemInput>::Inner<'static>: Send, <I as bevy::prelude::SystemInput>::Inner<'static>: Sync, <I as bevy::prelude::SystemInput>::Inner<'static>: Clone,
+    I: Send + Sync + SystemInput + 'static + Clone,
+    <I as SystemInput>::Inner<'static>: Send + Sync + Clone,
     O: Send + Sync + 'static,
 {
     move |world: &mut World| syscall(world, input.clone(), system.clone())
@@ -136,28 +139,38 @@ where
 pub trait WorldSyscallExt
 {
     /// See [`syscall`].
-    fn syscall<I, O, S, Marker>(&mut self, input:  <I as bevy::prelude::SystemInput>::Inner<'static>, system: S) -> O
+    fn syscall<I, O, S, Marker>(&mut self, input: <I as SystemInput>::Inner<'static>, system: S) -> O
     where
         I: Send + Sync + SystemInput + 'static,
         O: Send + Sync + 'static,
         S: IntoSystem<I, O, Marker> + Send + Sync + 'static;
 
     /// See [`syscall_with_validation`].
-    fn syscall_with_validation<I, O, S, Marker>(&mut self, input:  <I as bevy::prelude::SystemInput>::Inner<'static>, system: S, validation: fn(&mut World)) -> O
+    fn syscall_with_validation<I, O, S, Marker>(
+        &mut self,
+        input: <I as SystemInput>::Inner<'static>,
+        system: S,
+        validation: fn(&mut World)
+    ) -> O
     where
         I: Send + Sync + SystemInput + 'static,
         O: Send + Sync + 'static,
         S: IntoSystem<I, O, Marker> + Send + Sync + 'static;
 
     /// Similar to [`syscall`] except the system is not cached for reuse.
-    fn syscall_once<I, O, S, Marker>(&mut self, input:  <I as bevy::prelude::SystemInput>::Inner<'static>, system: S) -> O
+    fn syscall_once<I, O, S, Marker>(&mut self, input: <I as SystemInput>::Inner<'static>, system: S) -> O
     where
         I: Send + Sync + SystemInput + 'static,
         O: Send + Sync + 'static,
         S: IntoSystem<I, O, Marker> + Send + Sync + 'static;
 
     /// Similar to [`syscall_with_validation`] except the system is not cached for reuse.
-    fn syscall_once_with_validation<I, O, S, Marker>(&mut self, input:  <I as bevy::prelude::SystemInput>::Inner<'static>, system: S, validation: fn(&mut World)) -> O
+    fn syscall_once_with_validation<I, O, S, Marker>(
+        &mut self,
+        input: <I as SystemInput>::Inner<'static>,
+        system: S,
+        validation: fn(&mut World)
+    ) -> O
     where
         I: Send + Sync + SystemInput + 'static,
         O: Send + Sync + 'static,
@@ -166,7 +179,7 @@ pub trait WorldSyscallExt
 
 impl WorldSyscallExt for World
 {
-    fn syscall<I, O, S, Marker>(&mut self, input:  <I as bevy::prelude::SystemInput>::Inner<'static>, system: S) -> O
+    fn syscall<I, O, S, Marker>(&mut self, input: <I as SystemInput>::Inner<'static>, system: S) -> O
     where
         I: Send + Sync + SystemInput + 'static,
         O: Send + Sync + 'static,
@@ -175,7 +188,12 @@ impl WorldSyscallExt for World
         syscall(self, input, system)
     }
 
-    fn syscall_with_validation<I, O, S, Marker>(&mut self, input: <I as bevy::prelude::SystemInput>::Inner<'static>, system: S, validation: fn(&mut World)) -> O
+    fn syscall_with_validation<I, O, S, Marker>(
+        &mut self,
+        input: <I as SystemInput>::Inner<'static>,
+        system: S,
+        validation: fn(&mut World)
+    ) -> O
     where
         I: Send + Sync + SystemInput + 'static,
         O: Send + Sync + 'static,
@@ -184,7 +202,7 @@ impl WorldSyscallExt for World
         syscall_with_validation(self, input, system, validation)
     }
 
-    fn syscall_once<I, O, S, Marker>(&mut self, input: <I as bevy::prelude::SystemInput>::Inner<'static>, system: S) -> O
+    fn syscall_once<I, O, S, Marker>(&mut self, input: <I as SystemInput>::Inner<'static>, system: S) -> O
     where
         I: Send + Sync + SystemInput + 'static,
         O: Send + Sync + 'static,
@@ -195,7 +213,12 @@ impl WorldSyscallExt for World
         sys.run(input, self)
     }
 
-    fn syscall_once_with_validation<I, O, S, Marker>(&mut self, input: <I as bevy::prelude::SystemInput>::Inner<'static>, system: S, validation: fn(&mut World)) -> O
+    fn syscall_once_with_validation<I, O, S, Marker>(
+        &mut self,
+        input: <I as SystemInput>::Inner<'static>,
+        system: S,
+        validation: fn(&mut World)
+    ) -> O
     where
         I: Send + Sync + SystemInput + 'static,
         O: Send + Sync + 'static,
@@ -214,95 +237,131 @@ impl WorldSyscallExt for World
 pub trait CommandsSyscallExt
 {
     /// See [`syscall`].
-    fn syscall<I, S, Marker>(&mut self, input:  <I as bevy::prelude::SystemInput>::Inner<'static>, system: S)
+    fn syscall<I, R, S, Marker>(&mut self, input: <I as SystemInput>::Inner<'static>, system: S)
     where
-        I: Send + Sync + SystemInput + 'static, <I as bevy::prelude::SystemInput>::Inner<'static>: Send, <I as bevy::prelude::SystemInput>::Inner<'static>: Sync,
-        S: IntoSystem<I, (), Marker> + Send + Sync + 'static;
+        I: Send + Sync + SystemInput + 'static,
+        <I as SystemInput>::Inner<'static>: Send + Sync,
+        R: ReactorResult,
+        S: IntoSystem<I, R, Marker> + Send + Sync + 'static;
 
     /// See [`syscall_with_validation`].
-    fn syscall_with_validation<I, S, Marker>(&mut self, input:  <I as bevy::prelude::SystemInput>::Inner<'static>, system: S, validation: fn(&mut World))
+    fn syscall_with_validation<I, R, S, Marker>(&mut self, input: <I as SystemInput>::Inner<'static>, system: S, validation: fn(&mut World))
     where
-        I: Send + Sync + SystemInput + 'static, <I as bevy::prelude::SystemInput>::Inner<'static>: Send, <I as bevy::prelude::SystemInput>::Inner<'static>: Sync,
-        S: IntoSystem<I, (), Marker> + Send + Sync + 'static;
+        I: Send + Sync + SystemInput + 'static,
+        <I as SystemInput>::Inner<'static>: Send + Sync,
+        R: ReactorResult,
+        S: IntoSystem<I, R, Marker> + Send + Sync + 'static;
 
     /// Similar to [`syscall`] except the system is not cached for reuse.
-    fn syscall_once<I, S, Marker>(&mut self, input:  <I as bevy::prelude::SystemInput>::Inner<'static>, system: S)
+    fn syscall_once<I, R, S, Marker>(&mut self, input: <I as SystemInput>::Inner<'static>, system: S)
     where
-        I: Send + Sync + SystemInput + 'static, <I as bevy::prelude::SystemInput>::Inner<'static>: Send, <I as bevy::prelude::SystemInput>::Inner<'static>: Sync,
-        S: IntoSystem<I, (), Marker> + Send + Sync + 'static;
+        I: Send + Sync + SystemInput + 'static,
+        <I as SystemInput>::Inner<'static>: Send + Sync,
+        R: ReactorResult,
+        S: IntoSystem<I, R, Marker> + Send + Sync + 'static;
 
     /// Similar to [`syscall_with_validation`] except the system is not cached for reuse.
-    fn syscall_once_with_validation<I, S, Marker>(&mut self, input:  <I as bevy::prelude::SystemInput>::Inner<'static>, system: S, validation: fn(&mut World))
+    fn syscall_once_with_validation<I, R, S, Marker>(&mut self, input: <I as SystemInput>::Inner<'static>, system: S, validation: fn(&mut World))
     where
-        I: Send + Sync + SystemInput + 'static, <I as bevy::prelude::SystemInput>::Inner<'static>: Send, <I as bevy::prelude::SystemInput>::Inner<'static>: Sync,
-        S: IntoSystem<I, (), Marker> + Send + Sync + 'static;
+        I: Send + Sync + SystemInput + 'static,
+        <I as SystemInput>::Inner<'static>: Send + Sync,
+        R: ReactorResult,
+        S: IntoSystem<I, R, Marker> + Send + Sync + 'static;
 }
 
 impl CommandsSyscallExt for Commands<'_, '_>
 {
-    fn syscall<I, S, Marker>(&mut self, input:  <I as bevy::prelude::SystemInput>::Inner<'static>, system: S)
+    fn syscall<I, R, S, Marker>(&mut self, input: <I as SystemInput>::Inner<'static>, system: S)
     where
-        I: Send + Sync + SystemInput + 'static, <I as bevy::prelude::SystemInput>::Inner<'static>: Send, <I as bevy::prelude::SystemInput>::Inner<'static>: Sync,
-        S: IntoSystem<I, (), Marker> + Send + Sync + 'static
+        I: Send + Sync + SystemInput + 'static,
+        <I as SystemInput>::Inner<'static>: Send + Sync,
+        R: ReactorResult,
+        S: IntoSystem<I, R, Marker> + Send + Sync + 'static
     {
-        self.queue(move |world: &mut World| { world.syscall(input, system); });
+        self.queue(move |world: &mut World| {
+            let result = world.syscall(input, system);
+            result.handle(world);
+        });
     }
 
-    fn syscall_with_validation<I, S, Marker>(&mut self, input:  <I as bevy::prelude::SystemInput>::Inner<'static>, system: S, validation: fn(&mut World))
+    fn syscall_with_validation<I, R, S, Marker>(&mut self, input: <I as SystemInput>::Inner<'static>, system: S, validation: fn(&mut World))
     where
-        I: Send + Sync + SystemInput + 'static, <I as bevy::prelude::SystemInput>::Inner<'static>: Send, <I as bevy::prelude::SystemInput>::Inner<'static>: Sync,
-        S: IntoSystem<I, (), Marker> + Send + Sync + 'static
+        I: Send + Sync + SystemInput + 'static,
+        <I as SystemInput>::Inner<'static>: Send + Sync,
+        R: ReactorResult,
+        S: IntoSystem<I, R, Marker> + Send + Sync + 'static
     {
-        self.queue(move |world: &mut World| { world.syscall_with_validation(input, system, validation); });
+        self.queue(move |world: &mut World| {
+            let result = world.syscall_with_validation(input, system, validation);
+            result.handle(world);
+        });
     }
 
-    fn syscall_once<I, S, Marker>(&mut self, input:  <I as bevy::prelude::SystemInput>::Inner<'static>, system: S)
+    fn syscall_once<I, R, S, Marker>(&mut self, input: <I as SystemInput>::Inner<'static>, system: S)
     where
-        I: Send + Sync + SystemInput + 'static, <I as bevy::prelude::SystemInput>::Inner<'static>: Send, <I as bevy::prelude::SystemInput>::Inner<'static>: Sync,
-        S: IntoSystem<I, (), Marker> + Send + Sync + 'static
+        I: Send + Sync + SystemInput + 'static,
+        <I as SystemInput>::Inner<'static>: Send + Sync,
+        R: ReactorResult,
+        S: IntoSystem<I, R, Marker> + Send + Sync + 'static
     {
-        self.queue(move |world: &mut World| { world.syscall_once(input, system); });
+        self.queue(move |world: &mut World| {
+            let result = world.syscall_once(input, system);
+            result.handle(world);
+        });
     }
 
-    fn syscall_once_with_validation<I, S, Marker>(&mut self, input:  <I as bevy::prelude::SystemInput>::Inner<'static>, system: S, validation: fn(&mut World))
+    fn syscall_once_with_validation<I, R, S, Marker>(&mut self, input: <I as SystemInput>::Inner<'static>, system: S, validation: fn(&mut World))
     where
-        I: Send + Sync + SystemInput + 'static, <I as bevy::prelude::SystemInput>::Inner<'static>: Send, <I as bevy::prelude::SystemInput>::Inner<'static>: Sync,
-        S: IntoSystem<I, (), Marker> + Send + Sync + 'static
+        I: Send + Sync + SystemInput + 'static,
+        <I as SystemInput>::Inner<'static>: Send + Sync,
+        R: ReactorResult,
+        S: IntoSystem<I, R, Marker> + Send + Sync + 'static
     {
-        self.queue(move |world: &mut World| { world.syscall_once_with_validation(input, system, validation); });
+        self.queue(move |world: &mut World| {
+            let result = world.syscall_once_with_validation(input, system, validation);
+            result.handle(world);
+        });
     }
 }
 
 impl CommandsSyscallExt for EntityCommands<'_>
 {
-    fn syscall<I, S, Marker>(&mut self, input:  <I as bevy::prelude::SystemInput>::Inner<'static>, system: S)
+    fn syscall<I, R, S, Marker>(&mut self, input: <I as SystemInput>::Inner<'static>, system: S)
     where
-        I: Send + Sync + SystemInput + 'static, <I as bevy::prelude::SystemInput>::Inner<'static>: Send, <I as bevy::prelude::SystemInput>::Inner<'static>: Sync,
-        S: IntoSystem<I, (), Marker> + Send + Sync + 'static
+        I: Send + Sync + SystemInput + 'static,
+        <I as SystemInput>::Inner<'static>: Send + Sync,
+        R: ReactorResult,
+        S: IntoSystem<I, R, Marker> + Send + Sync + 'static
     {
         self.commands().syscall(input, system);
     }
 
-    fn syscall_with_validation<I, S, Marker>(&mut self, input:  <I as bevy::prelude::SystemInput>::Inner<'static>, system: S, validation: fn(&mut World))
+    fn syscall_with_validation<I, R, S, Marker>(&mut self, input: <I as SystemInput>::Inner<'static>, system: S, validation: fn(&mut World))
     where
-        I: Send + Sync + SystemInput + 'static, <I as bevy::prelude::SystemInput>::Inner<'static>: Send, <I as bevy::prelude::SystemInput>::Inner<'static>: Sync,
-        S: IntoSystem<I, (), Marker> + Send + Sync + 'static
+        I: Send + Sync + SystemInput + 'static,
+        <I as SystemInput>::Inner<'static>: Send + Sync,
+        R: ReactorResult,
+        S: IntoSystem<I, R, Marker> + Send + Sync + 'static
     {
         self.commands().syscall_with_validation(input, system, validation);
     }
 
-    fn syscall_once<I, S, Marker>(&mut self, input:  <I as bevy::prelude::SystemInput>::Inner<'static>, system: S)
+    fn syscall_once<I, R, S, Marker>(&mut self, input: <I as SystemInput>::Inner<'static>, system: S)
     where
-        I: Send + Sync + SystemInput + 'static, <I as bevy::prelude::SystemInput>::Inner<'static>: Send, <I as bevy::prelude::SystemInput>::Inner<'static>: Sync,
-        S: IntoSystem<I, (), Marker> + Send + Sync + 'static
+        I: Send + Sync + SystemInput + 'static,
+        <I as SystemInput>::Inner<'static>: Send + Sync,
+        R: ReactorResult,
+        S: IntoSystem<I, R, Marker> + Send + Sync + 'static
     {
         self.commands().syscall_once(input, system);
     }
 
-    fn syscall_once_with_validation<I, S, Marker>(&mut self, input:  <I as bevy::prelude::SystemInput>::Inner<'static>, system: S, validation: fn(&mut World))
+    fn syscall_once_with_validation<I, R, S, Marker>(&mut self, input: <I as SystemInput>::Inner<'static>, system: S, validation: fn(&mut World))
     where
-        I: Send + Sync + SystemInput + 'static, <I as bevy::prelude::SystemInput>::Inner<'static>: Send, <I as bevy::prelude::SystemInput>::Inner<'static>: Sync,
-        S: IntoSystem<I, (), Marker> + Send + Sync + 'static
+        I: Send + Sync + SystemInput + 'static,
+        <I as SystemInput>::Inner<'static>: Send + Sync,
+        R: ReactorResult,
+        S: IntoSystem<I, R, Marker> + Send + Sync + 'static
     {
         self.commands().syscall_once_with_validation(input, system, validation);
     }